+++ lib/databaseProvider.js
@@ -8,7 +8,6 @@
  */
 
 const database = process.require('lib/database.js');
-const conf = process.require('lib/conf.js');
 
 const getWithNormalizedKey = (obj, normalizedKey) => {
   for (const [key, value] of Object.entries(obj)) {
@@ -18,48 +17,64 @@ const getWithNormalizedKey = (obj, normalizedKey) => {
   }
 };
 
-/**
- * Fetches user from database.
- *
- * User attribute containing the login is defined by the userLoginAttribute property of the server configuration.
- *
- * @method getUser
- * @static
- * @params {String} login User login
- * @return {Object} The user as is
- */
-module.exports.getUser = (login) => {
-  const users = database.users;
-  const loginAttribute = conf.server.userLoginAttribute;
-  let fetchedUser;
-  let user = {attributes: {}};
+const deduplicate = (entries) => {
+  const unique = [];
+  entries.forEach((entry) => {
+    if (!unique.some((item) => item.dn === entry.dn)) {
+      unique.push(entry);
+    }
+  });
 
-  for (let user in users) {
-    if (users[user][loginAttribute] === login) {
-      fetchedUser = users[user];
-      break;
+  return unique;
+};
+
+const getIntersection = (results) => {
+  if (results.length < 1) {
+    return results;
+  }
+
+  const intersection = results[0];
+  for (const entries of results.slice(1)) {
+    for (const entry of entries) {
+      const idx = intersection.findIndex((item) => item.dn === entry.dn);
+      if (idx === -1) {
+        intersection.splice(idx, 1);
+      }
     }
   }
 
-  if (!fetchedUser) return null;
-
-  user.dn = fetchedUser.dn;
-  for (let propertyName in fetchedUser) {
-    if (propertyName !== 'dn')
-      user.attributes[propertyName] = fetchedUser[propertyName];
-  }
-
-  return user;
+  return intersection;
 };
 
-module.exports.getAllUsers = () => {
+const splitFilter = (filter) => {
+  const parts = [];
+
+  let parenCount = 0;
+  let start = 0;
+  Array.from(filter).forEach((char, i) => {
+    if (char === '(') {
+      parenCount++;
+    } else if (char === ')') {
+      parenCount--;
+    }
+
+    if (parenCount === 0) {
+      parts.push(filter.substring(start, i + 1));
+      start = i + 1;
+    }
+  });
+
+  return parts;
+};
+
+const getAllUsers = () => {
   return database.users.map((user) => {
     const {dn, ...attributes} = user;
     return {dn, attributes};
   });
 };
 
-module.exports.getUsers = (attributeName, searchValue) => {
+const getUsers = (attributeName, searchValue) => {
   return database.users
     .filter((user) => {
       const searchRegex = new RegExp(`^${searchValue.replace(/[.+?^${}()|[\]\\]/g, '\\$&').replace('*', '.*')}$`);
@@ -71,3 +86,33 @@ module.exports.getUsers = (attributeName, searchValue) => {
       return {dn, attributes};
     });
 };
+
+/**
+ * Searches the database.
+ *
+ * @method search
+ * @static
+ * @params {String} filter An LDAP search filter
+ * @return {Array<Object>} An array of user objects matching the filter
+ */
+const search = (filter) => {
+  const withoutParens = filter.substring(1, filter.length - 1);
+
+  if (withoutParens.startsWith('&')) {
+    return getIntersection(splitFilter(withoutParens.substring(1)).map(search));
+  } else if (withoutParens.startsWith('|')) {
+    return deduplicate(splitFilter(withoutParens.substring(1)).map(search).flat());
+  } else if (withoutParens.startsWith('!')) {
+    const filteredDns = search(withoutParens.substring(1)).map((user) => user.dn);
+    return getAllUsers().filter((user) => !filteredDns.includes(user.dn));
+  } else {
+    // we do this instead of just .split('=') because filters can have "=" in the values
+    const equalSignIndex = withoutParens.indexOf('=');
+    const attributeName = withoutParens.substring(0, equalSignIndex);
+    const attributeValue = withoutParens.substring(equalSignIndex + 1);
+
+    return getUsers(attributeName, attributeValue);
+  }
+};
+
+module.exports.search = search;
